added tabs and real ui to settings overlay component
