+++ src/renderer/Components/Settings/AboutSettingsComponent.tsx
@@ -1,0 +1,10 @@
+export function AboutSettingsComponent() {
+    return (
+        <div className="AboutSettingsWrapper">
+            <h3>About</h3>
+            <p>Path of Levelling 2 - Settings Overlay</p>
+            <p>Version 1.0.0</p>
+            <p>Created by Your Name</p>
+        </div>
+    );
+}
