+++ src/renderer/Components/Settings/GemSettingsComponent.tsx
@@ -1,0 +1,9 @@
+export function GemSettingsComponent() {
+    return (
+        <div className="GemsSettingsWrapper">
+            <h3>Gems Settings</h3>
+            <p>Configure your gem tracking and preferences here.</p>
+            {/* Add your gems settings UI here */}
+        </div>
+    );
+}
